--- Content.Server/Bed/Sleep/SleepingComponent.cs
@@ -1,9 +1,0 @@
-using Robust.Shared.GameStates;
-using Content.Shared.Bed.Sleep;
-
-namespace Content.Server.Bed.Sleep
-{
-    [NetworkedComponent, RegisterComponent]
-    public sealed class SleepingComponent : SharedSleepingComponent
-    {}
-}
