--- Content.Shared/Bed/Sleep/SharedSleepingComponent.cs
@@ -1,5 +1,0 @@
-namespace Content.Shared.Bed.Sleep
-{
-    public abstract class SharedSleepingComponent : Component
-    {}
-}
