+++ Content.Shared/Bed/Sleep/SleepingComponent.cs
@@ -1,0 +1,7 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Bed.Sleep;
+
+[NetworkedComponent, RegisterComponent]
+public sealed class SleepingComponent : Component
+{}
